mm: Add list based physical allocator functions
Free pages of physical address space are initialised into a single linked list.
Allocator keeps track of the head and tail of the list.

New pages are allocated from the head of the list. Head's next element becomes
then the new head.
Deallocated pages are appended to the tail of the list and become the new tail.

Thanks to that, both allocation and deallocation operate in constant time